refactored slicer logic
