+++ asset/package.json
@@ -4,7 +4,7 @@
     "main": "utils.js",
     "devDependencies": {
         "@types/bluebird": "^3.5.28",
-        "@types/got": "^9.6.7"
+        "@types/got": "^9.6.8"
     },
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1"
@@ -13,10 +13,10 @@
     "license": "MIT",
     "description": "",
     "dependencies": {
-        "@terascope/elasticsearch-api": "^2.1.13",
+        "@terascope/elasticsearch-api": "^2.2.0",
         "@terascope/error-parser": "^1.0.1",
         "@terascope/job-components": "^0.23.9",
-        "@terascope/teraslice-state-storage": "^0.7.10",
+        "@terascope/teraslice-state-storage": "^0.8.0",
         "bluebird": "^3.5.5",
         "datemath-parser": "^1.0.6",
         "got": "^9.6.0",
