+++ package.json
@@ -31,7 +31,7 @@
         "@terascope/eslint-config": "^0.1.6",
         "@terascope/job-components": "^0.23.9",
         "@types/jest": "^24.0.21",
-        "@types/node": "^12.12.3",
+        "@types/node": "^12.12.5",
         "eslint": "^6.6.0",
         "jest": "^24.8.0",
         "jest-extended": "^0.11.2",
@@ -39,6 +39,6 @@
         "teraslice-test-harness": "^0.8.12",
         "ts-jest": "^24.1.0",
         "ts-node": "^8.4.1",
-        "typescript": "^3.6.4"
+        "typescript": "^3.7.2"
     }
 }
